code cleanup

+++ conveyor-persistence/conveyor-persistence-core/src/main/java/com/aegisql/conveyor/persistence/converters/DoubleToBytesConverter.java
@@ -17,7 +17,7 @@ public class DoubleToBytesConverter implements ObjectToByteArrayConverter<Double
 			return null;
 		}
 		byte[] bytes = new byte[8];
-		ByteBuffer.wrap(bytes).putDouble(obj.doubleValue());
+		ByteBuffer.wrap(bytes).putDouble(obj);
 		return bytes;
 	}
