+++ conveyor-persistence/conveyor-persistence-core/src/main/java/com/aegisql/conveyor/persistence/converters/FloatToBytesConverter.java
@@ -17,7 +17,7 @@ public class FloatToBytesConverter implements ObjectToByteArrayConverter<Float> 
 			return null;
 		}
 		byte[] bytes = new byte[4];
-		ByteBuffer.wrap(bytes).putFloat(obj.floatValue());
+		ByteBuffer.wrap(bytes).putFloat(obj);
 		return bytes;
 	}
