+++ conveyor-persistence/conveyor-persistence-core/src/main/java/com/aegisql/conveyor/persistence/converters/ShortToBytesConverter.java
@@ -17,7 +17,7 @@ public class ShortToBytesConverter implements ObjectToByteArrayConverter<Short> 
 			return null;
 		}
 		byte[] bytes = new byte[2];
-		ByteBuffer.wrap(bytes).putShort(obj.shortValue());
+		ByteBuffer.wrap(bytes).putShort(obj);
 		return bytes;
 	}
