+++ conveyor-persistence/conveyor-persistence-core/src/main/java/com/aegisql/conveyor/persistence/converters/arrays/BytesToBytesConverter.java
@@ -16,7 +16,7 @@ public class BytesToBytesConverter implements ObjectArrayToByteArrayConverter<By
 		}
 		byte[] res = new byte[obj.length];
 		for(int i=0; i< obj.length; i++){
-			res[i] = obj[i].byteValue();
+			res[i] = obj[i];
 		}
 		return res;
 	}
