+++ conveyor-persistence/conveyor-persistence-core/src/main/java/com/aegisql/conveyor/persistence/converters/arrays/DoublesToBytesConverter.java
@@ -21,7 +21,7 @@ public class DoublesToBytesConverter implements ObjectArrayToByteArrayConverter<
 		ByteBuffer bb = ByteBuffer.wrap(res);
 		
 		for(int i = 0; i < obj.length; i++) {
-			bb.putDouble(8*i, obj[i].doubleValue());
+			bb.putDouble(8*i, obj[i]);
 		}
 		
 		return res;
